// src/pages/annotator/components/video-player/video-controls/video-controls.interface.ts
// INTEL CONFIDENTIAL
//
// Copyright (C) 2021 Intel Corporation
//
// This software and the related documents are Intel copyrighted materials, and your use of them is governed by
// the express license under which they were provided to you ("License"). Unless the License provides otherwise,
// you may not use, modify, copy, publish, distribute, disclose or transmit this software or the related documents
// without Intel's prior written permission.
//
// This software and the related documents are provided as is, with no express or implied warranties,
// other than those that are expressly stated in the License.

export interface VideoControls {
    canSelectPrevious: boolean;
    previous: () => void;
    canSelectNext: boolean;
    isPlaying: boolean;
    next: () => void;
    play: () => void;
    pause: () => void;
    goto: (frameNumber: number, showConfirmation?: boolean) => void;
}